refactor + add colleteral/futures market endpoints

+++ src/main/java/trade/wayruha/whitebit/dto/Time.java
@@ -2,9 +2,7 @@ package trade.wayruha.whitebit.dto;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class Time {
-    private BigInteger time;
+  private long time;
 }
